initial commit to move non-torch related functions out of test/run_test.py
